requirements.txt
# CrisisMapper Requirements
# AI-Powered Disaster Detection & Geospatial Analytics Platform

# Core AI/ML Dependencies
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.3.0
pillow>=9.5.0

# Computer Vision & Image Processing
albumentations>=1.3.0
imgaug>=0.4.0
imageio>=2.31.0
scikit-image>=0.21.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Geospatial Processing
gdal>=3.6.0
geopandas>=0.13.0
rasterio>=1.3.0
shapely>=2.0.0
pyproj>=3.5.0
fiona>=1.9.0
geopy>=2.3.0
folium>=0.14.0
leafmap>=0.20.0

# Data Processing & Analysis
pandas>=2.0.0
xarray>=2023.1.0
dask>=2023.1.0
h5py>=3.8.0
zarr>=2.15.0
netCDF4>=1.6.0

# Web Framework & API
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
pydantic>=2.0.0
python-multipart>=0.0.6
httpx>=0.24.0
requests>=2.31.0

# Dashboard & Visualization
streamlit>=1.25.0
plotly>=5.15.0
altair>=5.1.0
bokeh>=3.2.0
dash>=2.11.0
dash-bootstrap-components>=1.4.0

# Database & Storage
sqlalchemy>=2.0.0
alembic>=1.11.0
psycopg2-binary>=2.9.0
redis>=4.6.0
pymongo>=4.4.0

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
cryptography>=41.0.0

# Monitoring & Logging
prometheus-client>=0.17.0
structlog>=23.1.0
loguru>=0.7.0
sentry-sdk[fastapi]>=1.28.0

# Testing & Quality
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0
black>=23.7.0
flake8>=6.0.0
mypy>=1.5.0
isort>=5.12.0
pre-commit>=3.3.0

# Development & Utilities
jupyter>=1.0.0
jupyterlab>=4.0.0
ipywidgets>=8.0.0
tqdm>=4.65.0
click>=8.1.0
rich>=13.4.0
typer>=0.9.0

# Cloud & Deployment
boto3>=1.28.0
google-cloud-storage>=2.10.0
azure-storage-blob>=12.17.0
docker>=6.1.0
kubernetes>=27.2.0

# Scientific Computing
numba>=0.57.0
numba-ml>=0.1.0
cupy-cuda11x>=12.0.0  # For CUDA 11.x
cupy-cuda12x>=12.0.0  # For CUDA 12.x

# Additional ML Libraries
transformers>=4.30.0
timm>=0.9.0
segmentation-models-pytorch>=0.3.0
pytorch-lightning>=2.0.0
wandb>=0.15.0
tensorboard>=2.13.0

# Geospatial Visualization
cartopy>=0.21.0
contextily>=1.4.0
mplleaflet>=0.0.5
keplergl>=0.3.0

# Data Validation & Serialization
marshmallow>=3.20.0
cerberus>=1.3.4
jsonschema>=4.19.0

# Configuration Management
pyyaml>=6.0
toml>=0.10.2
python-dotenv>=1.0.0
hydra-core>=1.3.0

# Async & Concurrency
aiofiles>=23.2.0
asyncio-mqtt>=0.13.0
aioredis>=2.0.0

# Image Processing & Computer Vision
imagehash>=4.3.1
face-recognition>=1.3.0
dlib>=19.24.0
mediapipe>=0.10.0

# Geospatial Data Sources
sentinelsat>=1.2.0
landsatxplore>=0.7.0
modis>=0.1.0
planet>=2.0.0

# Performance & Optimization
numba>=0.57.0
cython>=3.0.0
numba-ml>=0.1.0

# Documentation
sphinx>=7.0.0
sphinx-rtd-theme>=1.3.0
mkdocs>=1.5.0
mkdocs-material>=9.1.0

# Additional Utilities
python-dateutil>=2.8.2
pytz>=2023.3
schedule>=1.2.0
APScheduler>=3.10.0
celery>=5.3.0

# Environment Management
conda>=23.5.0
pip-tools>=7.0.0
pipenv>=2023.7.0

# Version Control
gitpython>=3.1.0
dulwich>=0.21.0

# Network & Communication
websockets>=11.0.0
aiohttp>=3.8.0
httpx>=0.24.0

# File Processing
openpyxl>=3.1.0
xlsxwriter>=3.1.0
pypdf2>=3.0.0
python-docx>=0.8.11

# Time Series
statsmodels>=0.14.0
prophet>=1.1.0
arch>=6.2.0

# Additional Scientific Libraries
sympy>=1.12.0
networkx>=3.1.0
igraph>=0.10.0

# Geospatial Analysis
osmnx>=1.6.0
momepy>=0.6.0
urbanaccess>=0.2.0

# Machine Learning Extensions
optuna>=3.2.0
hyperopt>=0.2.7
ray[tune]>=2.6.0

# Data Visualization Extensions
bokeh>=3.2.0
holoviews>=1.16.0
panel>=1.2.0

# Additional Dependencies
tensorflow>=2.13.0  # Optional: for TensorFlow models
keras>=2.13.0       # Optional: for Keras models
xgboost>=1.7.0      # Optional: for XGBoost models
lightgbm>=4.0.0     # Optional: for LightGBM models
catboost>=1.2.0     # Optional: for CatBoost models